feat: 新しいストーリーを追加し、不要なファイルを削除
- hit-numbers.stories.tsを新規作成し、HitNumbersコンポーネントのストーリーを追加
- contorol-buttons.stores.tsを新規作成し、ControlButtonsコンポーネントのストーリーを追加
- Button.stories.tsのタイトルを変更
- 不要なCSSファイル（header.css、button.css、page.css）を削除
- 不要なストーリーファイル（Header.stories.ts、Button.tsx、Page.stories.ts）を削除
- 不要なSVGファイル（youtube.svg、tutorials.svg）を削除
