+++ application/src/stories/Button.stories.ts
@@ -2,9 +2,8 @@ import type {Meta, StoryObj} from "@storybook/react"
 import {fn} from "@storybook/test"
 import {Button} from "@/components/ui/button"
 
-// More on how to set up stories at: https://storybook.js.org/docs/writing-stories#default-export
 const meta: Meta<typeof Button> = {
-	title: "Buttons",
+	title: "LeftSide/Buttons/Button",
 	component: Button,
 	parameters: {
 		// Optional parameter to center the component in the Canvas. More info: https://storybook.js.org/docs/configure/story-layout
@@ -20,7 +19,6 @@ const meta: Meta<typeof Button> = {
 export default meta
 type Story = StoryObj<typeof meta>
 
-// More on writing stories with args: https://storybook.js.org/docs/writing-stories/args
 export const Start: Story = {
 	args: {
 		asChild: false,
