--- application/src/stories/Header.stories.ts
@@ -1,33 +1,0 @@
-import type {Meta, StoryObj} from "@storybook/react"
-import {fn} from "@storybook/test"
-
-import {Header} from "./Header"
-
-const meta: Meta<typeof Header> = {
-	title: "Example/Header",
-	component: Header,
-	// This component will have an automatically generated Autodocs entry: https://storybook.js.org/docs/writing-docs/autodocs
-	tags: ["autodocs"],
-	parameters: {
-		// More on how to position stories at: https://storybook.js.org/docs/configure/story-layout
-		layout: "fullscreen",
-	},
-	args: {
-		onLogin: fn(),
-		onLogout: fn(),
-		onCreateAccount: fn(),
-	},
-} satisfies Meta<typeof Header>
-
-export default meta
-type Story = StoryObj<typeof meta>
-
-export const LoggedIn: Story = {
-	args: {
-		user: {
-			name: "Jane Doe",
-		},
-	},
-}
-
-export const LoggedOut: Story = {}
